% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_data.R
\name{descr_NA}
\alias{descr_NA}
\title{Title Create descriptive statistics for incomplete data}
\usage{
descr_NA(d)
}
\arguments{
\item{d}{An incomplete dataset of class dataframe}
}
\value{
A datatable object with descriptives
}
\description{
Title Create descriptive statistics for incomplete data
}
