% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{dummy_plot}
\alias{dummy_plot}
\title{server utility}
\usage{
dummy_plot()
}
\arguments{
\item{Internal}{parameters for the utility.}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A shiny utility
}
