% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{\%nin\%}
\alias{\%nin\%}
\title{server utility}
\usage{
... \%nin\% NA
}
\arguments{
\item{Internal}{parameters for the utility.}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A shiny utility
}
