% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{no_br}
\alias{no_br}
\title{ui utility}
\usage{
no_br()
}
\arguments{
\item{Internal}{parameters for the utility.}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A shiny utility
}
