% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{plot_bw}
\alias{plot_bw}
\title{Title Plot boxplot/faceted barplot of imputed data}
\usage{
plot_bw(imp, x)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}
}
\value{
A ggplot object to replace mice::bwplot()
}
\description{
Title Plot boxplot/faceted barplot of imputed data
}
