% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_02_model.R
\name{plot_flux}
\alias{plot_flux}
\title{Title Plot in- and outflux of an incomplete dataset}
\usage{
plot_flux(dat)
}
\arguments{
\item{dat}{An incomplete dataset of class dataframe}
}
\value{
A ggplot object with influx plotted against outflux
}
\description{
Title Plot in- and outflux of an incomplete dataset
}
