% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{plot_imps}
\alias{plot_imps}
\title{Title Preprocess imputations and initialize plot}
\usage{
plot_imps(imp, x, y = NULL)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}

\item{y}{An optional second variable to plot}
}
\value{
A ggplot object with imputed data
}
\description{
Title Preprocess imputations and initialize plot
}
