% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_02_model.R
\name{plot_md_pattern}
\alias{plot_md_pattern}
\title{Title Plot missing data pattern of incomplete data}
\usage{
plot_md_pattern(dat)
}
\arguments{
\item{dat}{An incomplete dataset of class dataframe}
}
\value{
A ggplot object visualizing the missing data pattern matrix
}
\description{
Title Plot missing data pattern of incomplete data
}
