% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_missingness_utils_server.R
\name{plot_NA_scatter}
\alias{plot_NA_scatter}
\title{Title}
\usage{
plot_NA_scatter(dat, x, y)
}
\arguments{
\item{dat}{An incomplete dataset of class dataframe}

\item{x}{A variable to plot}

\item{y}{A second variable to plot}
}
\value{

}
\description{
Title
}
