% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputationmodel_utils_server.R
\name{plot_pred_matrix}
\alias{plot_pred_matrix}
\title{Title}
\usage{
plot_pred_matrix(pred)
}
\arguments{
\item{d}{An incomplete dataset of class dataframe, a multiply imputed data set of class mids, or a predictor matrix}
}
\value{

}
\description{
Title
}
