% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputationmodel_utils_server.R
\name{plot_rhat}
\alias{plot_rhat}
\title{Title}
\usage{
plot_rhat(imp, x, theta = "means")
}
\arguments{
\item{imp}{A multiply imputed data set (mids) object}

\item{x}{A variable to plot}

\item{theta}{The parameter to plot (either "means" for chain means, or "vars" for chain variances)}
}
\value{

}
\description{
Title
}
