% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{plot_strip}
\alias{plot_strip}
\title{Title Plot stripplot of imputed data}
\usage{
plot_strip(imp, x)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}
}
\value{
A ggplot object to replace mice::stripplot
}
\description{
Title Plot stripplot of imputed data
}
