% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeddata_utils_server.R
\name{plot_xy}
\alias{plot_xy}
\title{Title}
\usage{
plot_xy(imp, x, y)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}

\item{y}{A second variable to plot}
}
\value{

}
\description{
Title
}
