% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_03_convergence.R
\name{preprocess_thetas}
\alias{preprocess_thetas}
\title{Title Preprocess imputation chains (i.e., chain means, chain variances) for traceplot}
\usage{
preprocess_thetas(imp)
}
\arguments{
\item{imp}{A multiply imputed data set (mids) object}
}
\value{
A dataframe with tidy convergence parameters for all variables
}
\description{
Title Preprocess imputation chains (i.e., chain means, chain variances) for traceplot
}
