% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_data.R
\name{tab_NA}
\alias{tab_NA}
\title{Title Tabulate an incomplete dataset and highlight the NAs}
\usage{
tab_NA(d)
}
\arguments{
\item{d}{An incomplete dataset of class dataframe}
}
\value{
A datatable object with the NA cells highlighted
}
\description{
Title Tabulate an incomplete dataset and highlight the NAs
}
