% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputationmodel_utils_server.R
\name{trace_one_variable}
\alias{trace_one_variable}
\title{Title}
\usage{
trace_one_variable(d, x)
}
\arguments{
\item{d}{A dataframe with preprocessed imputation chains generated via the function `preprocess_thetas()`}

\item{x}{A variable to plot}
}
\value{

}
\description{
Title
}
