% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeddata_utils_server.R
\name{plot_bw}
\alias{plot_bw}
\title{Title}
\usage{
plot_bw(imp, x)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}
}
\value{

}
\description{
Title
}
