% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{plot_dens}
\alias{plot_dens}
\title{Title Plot density of imputed data}
\usage{
plot_dens(imp, x)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}
}
\value{
A ggplot object to replace mice::densityplot
}
\description{
Title Plot density of imputed data
}
