% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_data.R
\name{plot_NA_scatter}
\alias{plot_NA_scatter}
\title{Title Create a scatterplot with NAs plotted in the margins}
\usage{
plot_NA_scatter(dat, x, y)
}
\arguments{
\item{dat}{An incomplete dataset of class dataframe}

\item{x}{A variable to plot}

\item{y}{A second variable to plot}
}
\value{
A ggplot object with NAs plotted in the margins
}
\description{
Title Create a scatterplot with NAs plotted in the margins
}
