% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_02_model.R
\name{plot_pred_matrix}
\alias{plot_pred_matrix}
\title{Title Plot the predictor matrix for the imputation model}
\usage{
plot_pred_matrix(pred)
}
\arguments{
\item{d}{A predictor matrix obtained from a mids object or using mice::quickpred()}
}
\value{
A ggplot object visualizing the predictor matrix
}
\description{
Title Plot the predictor matrix for the imputation model
}
