% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputationmodel_utils_server.R
\name{preprocess_thetas}
\alias{preprocess_thetas}
\title{Title}
\usage{
preprocess_thetas(imp)
}
\arguments{
\item{imp}{A multiply imputed data set (mids) object}
}
\value{

}
\description{
Title
}
